+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Settings/CarlaSettings.cpp
@@ -162,7 +162,7 @@ void UCarlaSettings::LoadSettings()
     FString RmwValue;
     if (FParse::Value(FCommandLine::Get(), TEXT("-rmw="), RmwValue))
     {
-      DDSMiddlewareName = RmwValue.ToLower();
+      MiddlewareName = RmwValue.ToLower();
     }
   }
 }